Update acl.xml - change configuration Move button into Button folder Rename BackButton.php Move button into Button folder Rename DeleteButton.php Update Edit.php - add ADMIN_RESOURCE Rename GenericButton.php Move button into Button folder Update menu.xml - change configuration Update RestaurantActions.php - remove delete logic Update restaurants_restaurant_edit.xml - clear configuration Update restaurants_restaurant_form.xml - change configuration
