+++ app/code/Stoyanov/Restaurant/Block/Adminhtml/Restaurant/Edit/Button/Back.php
@@ -1,11 +1,11 @@
 <?php
 declare(strict_types=1);
 
-namespace Stoyanov\Restaurant\Block\Adminhtml\Page\Edit;
+namespace Stoyanov\Restaurant\Block\Adminhtml\Restaurant\Edit\Button;
 
 use Magento\Framework\View\Element\UiComponent\Control\ButtonProviderInterface;
 
-class BackButton extends GenericButton implements ButtonProviderInterface
+class Back extends Generic implements ButtonProviderInterface
 {
     /**
      * @return array
