+++ app/code/Stoyanov/Restaurant/etc/acl.xml
@@ -2,17 +2,14 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Acl/etc/acl.xsd">
     <acl>
         <resources>
-            <!-- Root ACL for Magento admin -->
             <resource id="Magento_Backend::admin">
-                <!-- Our module's root ACL -->
-                <resource id="Stoyanov_Restaurant::restaurant" title="Restaurants" sortOrder="10">
-                    <!-- Sub-ACL for view -->
-                    <resource id="Stoyanov_Restaurant::restaurant_view" title="View Restaurants" sortOrder="15"/>
-
-                    <!-- Sub-ACL for edit -->
-                    <resource id="Stoyanov_Restaurant::restaurant_edit" title="Edit Restaurants" sortOrder="20"/>
-                    <resource id="Stoyanov_Restaurant::restaurant_save" title="Save Restaurant" sortOrder="30"/>
-                    <resource id="Stoyanov_Restaurant::restaurant_delete" title="Delete Restaurant" sortOrder="40"/>
+                <resource id="Magento_Backend::content">
+                    <resource id="Stoyanov_Restaurant::stoyanov" title="Restaurant" translate="title" sortOrder="900">
+                        <resource id="Stoyanov_Restaurant::restaurant" title="FAQs" translate="title">
+                            <resource id="Stoyanov_Restaurant::restaurant_save" title="Save" translate="title"/>
+                            <resource id="Stoyanov_Restaurant::restaurant_delete" title="Delete" translate="title"/>
+                        </resource>
+                    </resource>
                 </resource>
             </resource>
         </resources>
