+++ app/code/Stoyanov/Restaurant/view/adminhtml/layout/restaurants_restaurant_edit.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0"?>
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <update handle="styles"/>
-    <update handle="editor"/>
     <body>
         <referenceContainer name="content">
             <uiComponent name="restaurants_restaurant_form"/>
